+++ tests/testthat/test_average_vimp_regression.R
@@ -1,9 +1,9 @@
-context("test_average_vimp_regression.R")
+context("Test averaging of difference nonparametric R^2 values")
 
 ## load required functions and packages
 library("testthat")
 library("SuperLearner")
-library("xgboost")
+library("gam")
 library("vimp")
 
 ## generate the data
@@ -18,7 +18,7 @@ y <- (x[,1]/5)^2*(x[,1]+7)/5 + (x[,2]/3)^2 + rnorm(n, 0, 1)
 samp <- sample(1:n, n/2, replace = FALSE)
 
 ## set up a library for SuperLearner
-learners <- "SL.xgboost"
+learners <- c("SL.step", "SL.gam", "SL.mean")
 
 ## fit the data with all covariates
 full_fit_1 <- SuperLearner(Y = y[samp], X = x[samp, ], SL.library = learners)
@@ -41,4 +41,6 @@ test_that("Test averaging of ANOVA-based variable importance", {
 
   est <- average_vim(est_1, est_2)
   expect_equal(est$est, (500/729)/(1 + 2497/7875 + 500/729), tolerance = 0.1)
+  expect_length(est$mat, 4)
+  expect_output(print(est), "\n\nVariable importance estimates:\n", fixed = TRUE)
 })